CF rework.
Numinous won't decode by line always now.

Now Numinous will wait for \n if its Evaluation task (perhaps RODATA)

If its if, else if or anything that has evaluation before braces. (or loop) then it will exit on capturing all it should for that step.

Assembly instructions are yet a little complex.

Rules on what to report as error may change from here.

Due to the major infrastructural change, the code is expected to perform awfully as is.

Will redesign extensively.