Engineered Lexer new StringIndex/Length relationships
When the Lexer was simple and linear the index or length of the last token was all that was important.

Those attributes belong to error report only, the user cares about what defines were passed not what they became.

So we report  based on the source code only, storing source references to the generated tokens from the defines.

Will need serious testing.