New definition system + Linker concept
Since we need to evaluate as much as we can regardless of undefined presence in a definition we accept that we can track all undefined symbols.

We can create those symbols and attach their dependants like before, but the magic is this time we only need to broadcast the name of the object thats been lazily defined.

Undefined tokens being able to track their undefined components while knowing how many are left undefined at any point should optimise this and generally create cleaner code.

Also we want a linker.