No more domain specified toml parsing
It just wasn't capable enough, so I'm writing the logic myself.

Because we ordered the entries it should be safe to go as we have. the new solution is just to convert the toml  data into our nice class layout.

Once this is tested, we can see if Linker Creation is usable